packages/wsjrdp2027/pyproject.toml
[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

[project]
name = "wsjrdp2027"
version = "0.0.1"
description = "Helpers for wsjrdp_scripts"
readme = "README.md"
authors = [
    { name = "David Fritzsche", email = "david.fritzsche@worldscoutjamboree.de" }
]
requires-python = ">=3.13"
dependencies = [
    "atlassian-python-api>=4",
    "python-editor",
    "setuptools",  # for python-editor
    "html2text>=2025.4.15",
    "humanfriendly>=10.0",
    "iso4217>=1.14.20250512",
    "jinja2>=3.1.6",
    "lxml",
    "pandas>=2.3.3",
    "paramiko<4",
    "psycopg[binary]>=3.3.2",
    "pyyaml>=6.0.2",
    "saxonche>=12.9.0",
    "schwifty>=2025.9.0",
    "sepaxml>=2.7.0",
    "sshtunnel>=0.4.0",
    "typst>=0.14",
    "xlsxwriter>=3.2.5",
    "dictdiffer>=0.9.0",
]

[dependency-groups]
pytest = [
    "pytest>=9.0.2",
    "time-machine>=3.2.0",
]
type-checking = [
    "mypy",
    "pandas-stubs>=2.3.3.251201",
    "ty",
    "types-humanfriendly>=10.0.1.20250319",
    "types-lxml",
]
